# vtprotobuf: Unmarshal uses pooled objects