@dmathieu 2023-07-25
Diff between the two commits in the fork: https://github.com/carsonip/vtprotobuf/compare/22cdafcdd35e..2cf3150fde0e