@dmathieu 2023-07-25
Won't this change be hard to get accepted in vtprotobuf upstream? Doesn't this mean we'll end up with a fork indefinitely?