@carsonip 2023-07-25
I'm also a little hesitant about this as well, and yes that means we'll end up with a fork.

But I believe upstream will go down the route of pooling the backing arrays which contain pointers to objects in the heap. e.g. for a `var s []*KeyedTransactionMetrics`, when it is returned to the pool, all the KeyedTransactionMetrics are reset, and the slice is only resliced, i.e. `s = s[:0]`. Upon reuse, it is resliced, i.e. `s = s[:len(s)+1]`, and then s[0] will contain an empty KeyedTransactionMetrics for reuse without alloc.

Currently for our usage, the upstream implementation may not be very efficient for us, since we try to obtain individual elements `Keyed*` from VTPool while the pool is never populated and it will result in allocs.

I'm happy to put this back into draft and wait for the code to stabilize a bit before we pick a route. It is micro-optimization anyway.